+++ CollegeManagement/AdvancedSession/ParallelManagement.cs
@@ -58,10 +58,15 @@ namespace CollegeManagement.AdvancedSession
 
         private void iconButton1_Click(object sender, EventArgs e)
         {
-            objCore.DeleteParallelSession(ConsecutiveID);
-            gvRemove.Visible = false;
-            iconButton1.Visible = false;
-            loadData();
+            DialogResult result = MessageBox.Show("Are you sure you want to Remove?", "Confirm Remove", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+
+            if (result.ToString().Equals("Yes"))
+            {
+                objCore.DeleteParallelSession(ConsecutiveID);
+                gvRemove.Visible = false;
+                iconButton1.Visible = false;
+                loadData();
+            }            
         }
     }
 }
